Added working leaderboard, and points
Using a demo add function, when a picture is added, the picture will be saved to the users private database. The user will gain 100 points and it will reflect on the main page. The leaderboard tab will also be updated. The refreshable list will show all users with points in a descending order. Adding points solution may not work verywell long term as multipliers and other features are added.